+++ zink/src/primitives/numeric.rs
@@ -28,42 +28,6 @@ macro_rules! local_revert {
     };
 }
 
-macro_rules! impl_numeric {
-    ($($t:ty, $addmod_fn:ident, $mulmod_fn:ident);* $(;)?) => {
-        $(
-            impl Numeric for $t {
-                #[inline(always)]
-                fn addmod(self, other: Self, n: Self) -> Self {
-                    #[cfg(target_arch = "wasm32")]
-                    unsafe { ffi::asm::$addmod_fn(n, other, self) }
-                    #[cfg(not(target_arch = "wasm32"))]
-                    ffi::asm::asm::$addmod_fn(n, other, self)
-                }
-                #[inline(always)]
-                fn mulmod(self, other: Self, n: Self) -> Self {
-                    #[cfg(target_arch = "wasm32")]
-                    unsafe { ffi::asm::$mulmod_fn(n, other, self) }
-                    #[cfg(not(target_arch = "wasm32"))]
-                    ffi::asm::asm::$mulmod_fn(n, other, self)
-                }
-            }
-        )*
-    };
-    // Special case for U256
-    (U256, $addmod_fn:ident, $mulmod_fn:ident) => {
-        impl Numeric for U256 {
-            #[inline(always)]
-            fn addmod(self, other: Self, n: Self) -> Self {
-                unsafe { ffi::$addmod_fn(n, other, self) }
-            }
-            #[inline(always)]
-            fn mulmod(self, other: Self, n: Self) -> Self {
-                unsafe { ffi::$mulmod_fn(n, other, self) }
-            }
-        }
-    };
-}
-
 // Signed types (i8, i16, i32, i64)
 macro_rules! impl_safe_numeric_signed {
     ($($t:ty);* $(;)?) => {
@@ -214,17 +178,6 @@ impl SafeNumeric for U256 {
         }
         unsafe { ffi::u256_div(self, rhs) }
     }
-}
-
-impl_numeric! {
-    i8, addmod_i8, mulmod_i8;
-    u8, addmod_u8, mulmod_u8;
-    i16, addmod_i16, mulmod_i16;
-    u16, addmod_u16, mulmod_u16;
-    i32, addmod_i32, mulmod_i32;
-    u32, addmod_u32, mulmod_u32;
-    i64, addmod_i64, mulmod_i64;
-    u64, addmod_u64, mulmod_u64;
 }
 
 impl_safe_numeric_signed! {
